Refactor CNodeHealthManager w secondary grace period
